[*] CharacterInfo requests moved to store
